requirements.txt
langchain-community==0.3.31
langchain-openai==0.3.35
langchain-text-splitters==0.3.9
faiss-cpu==1.12.0
python-dotenv==1.2.1